start admin event viewer and dashboard
