--- txcdr-client/app/(auth)/event/AdminPage.tsx
@@ -1,5 +1,0 @@
-import { PageProps } from "./types";
-
-export function AdminPage(props: PageProps) {
-  return <></>;
-}
